requirements.txt
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.2
pillow==11.0.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
typing_extensions==4.12.2
tzdata==2025.2
